Made InstanceIdClient into an interface
